move section conditions
