+++ src/packages/core/extension-registry/conditions/manifests.ts
@@ -1,10 +1,5 @@
 import type { ManifestTypes } from '../models/index.js';
 import { manifest as multipleAppLanguagesConditionManifest } from './multiple-app-languages.condition.js';
-import { manifest as sectionAliasConditionManifest } from './section-alias.condition.js';
 import { manifest as switchConditionManifest } from './switch.condition.js';
 
-export const manifests: Array<ManifestTypes> = [
-	multipleAppLanguagesConditionManifest,
-	sectionAliasConditionManifest,
-	switchConditionManifest,
-];
+export const manifests: Array<ManifestTypes> = [multipleAppLanguagesConditionManifest, switchConditionManifest];
