+++ src/packages/core/extension-registry/conditions/types.ts
@@ -1,7 +1,6 @@
 import type { CollectionAliasConditionConfig } from '../../collection/collection-alias.manifest.js';
 import type { CollectionBulkActionPermissionConditionConfig } from '../../collection/collection-bulk-action-permission.manifest.js';
 import type { UmbSectionUserPermissionConditionConfig } from '../../section/conditions/index.js';
-import type { SectionAliasConditionConfig } from './section-alias.condition.js';
 import type { SwitchConditionConfig } from './switch.condition.js';
 import type { UmbMultipleAppLanguageConditionConfig } from './multiple-app-languages.condition.js';
 import type { UmbConditionConfigBase } from '@umbraco-cms/backoffice/extension-api';
@@ -19,7 +18,6 @@ export type ConditionTypes =
 	| BlockWorkspaceHasSettingsConditionConfig
 	| CollectionAliasConditionConfig
 	| CollectionBulkActionPermissionConditionConfig
-	| SectionAliasConditionConfig
 	| SwitchConditionConfig
 	| UmbConditionConfigBase
 	| UmbDocumentUserPermissionConditionConfig
