+++ src/packages/core/section/manifests.ts
@@ -1,4 +1,4 @@
-import { UmbSectionUserPermissionCondition } from './conditions/section-user-permission.condition.js';
+import { manifests as sectionUserPermissionConditionManifests } from './conditions/manifests.js';
 import { manifests as repositoryManifests } from './repository/manifests.js';
 import type { ManifestTypes } from '@umbraco-cms/backoffice/extension-registry';
 
@@ -9,11 +9,6 @@ export const manifests: Array<ManifestTypes> = [
 		name: 'Section Picker Modal',
 		element: () => import('./section-picker-modal/section-picker-modal.element.js'),
 	},
-	{
-		type: 'condition',
-		name: 'Section User Permission Condition',
-		alias: 'Umb.Condition.SectionUserPermission',
-		api: UmbSectionUserPermissionCondition,
-	},
+	...sectionUserPermissionConditionManifests,
 	...repositoryManifests,
 ];
